requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
SQLAlchemy==2.0.23
WooCommerce==3.0.0
requests==2.31.0
python-dotenv==1.0.0
Pillow==11.1.0
rich==13.7.0
python-slugify==8.0.1
fdb==2.0.2
